templates/paper-to-api/requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pyyaml>=6.0

# ML — uncomment what you need
torch>=2.1.0
# transformers>=4.35.0
# diffusers>=0.24.0
# accelerate>=0.25.0

# Monitoring
# prometheus-fastapi-instrumentator>=6.0.0

# Testing
pytest>=7.4.0
httpx>=0.25.0
